requirements.txt
opencv-python
cmake
pyyaml
streamlit==1.22.0
cryptography
numpy
Pillow

# Fork of face_recognition using prebuilt dlib library instead
face_recognition @ git+https://github.com/thetoby9944/face_recognition

